+++ src/main/java/com/tstepnik/planner/domain/task/Importance.java
@@ -1,4 +1,4 @@
-package com.tstepnik.planner.domain;
+package com.tstepnik.planner.domain.task;
 
 public enum Importance {
     IMPORTANT,
