+++ src/main/java/com/tstepnik/planner/domain/user/UserRole.java
@@ -1,0 +1,25 @@
+package com.tstepnik.planner.domain.user;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import javax.persistence.*;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+public class UserRole {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Enumerated(EnumType.STRING)
+    private Role role;
+
+    public UserRole(Role role) {
+        this.role = role;
+    }
+}
