+++ src/main/java/com/tstepnik/planner/repository/TaskRepository.java
@@ -1,8 +1,7 @@
 package com.tstepnik.planner.repository;
 
-import com.tstepnik.planner.domain.Task;
+import com.tstepnik.planner.domain.task.Task;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
