+++ src/main/java/com/tstepnik/planner/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.tstepnik.planner.repository;
 
-import com.tstepnik.planner.domain.User;
+import com.tstepnik.planner.domain.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
