+++ src/main/java/com/tstepnik/planner/service/TaskService.java
@@ -1,12 +1,16 @@
 package com.tstepnik.planner.service;
 
-import com.tstepnik.planner.domain.*;
+import com.tstepnik.planner.domain.task.Importance;
+import com.tstepnik.planner.domain.task.Task;
+import com.tstepnik.planner.domain.user.User;
 import com.tstepnik.planner.exceptions.TaskExpiredException;
 import com.tstepnik.planner.repository.TaskRepository;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
-import java.time.*;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.LocalTime;
 import java.time.temporal.ChronoUnit;
 import java.util.List;
 
@@ -44,13 +48,13 @@ public class TaskService {
 
     public Task addTask(Task task) {
         User user = authService.getLoggedUser();
-        ZonedDateTime now = ZonedDateTime.now(ZoneId.of("UTC")).truncatedTo(ChronoUnit.SECONDS);
+        LocalDateTime now = LocalDateTime.now().truncatedTo(ChronoUnit.SECONDS);
         task.setCreationDate(now);
         if (task.getImportance() == null) {
             task.setImportance(DEFAULT_IMPORTANCE);
         }
         if (task.getPlannedFor() == null) {
-            ZonedDateTime defaultPlannedTime = ZonedDateTime.of(LocalDate.now(ZoneId.of("UTC")).plusDays(1), LocalTime.MIDNIGHT, ZoneId.of("UTC"));
+            LocalDateTime defaultPlannedTime = LocalDateTime.of(LocalDate.now().plusDays(1), LocalTime.MIDNIGHT);
             task.setPlannedFor(defaultPlannedTime);
         }
         task.setUserId(user.getId());
@@ -62,12 +66,12 @@ public class TaskService {
         Task checkedTask = taskRepository.getOne(taskId);
         if (!checkedTask.getUserId().equals(user.getId())) {
             throw new AccessDeniedException("User with login " + user.getId() + " cannot edit task with id " + taskId);
-        } else if (checkedTask.getPlannedFor().isBefore(ZonedDateTime.now(ZoneId.of("UTC")))) {
-            throw new TaskExpiredException("Time for finish this task had expired at: " + checkedTask.getPlannedFor());
+        } else if (checkedTask.getPlannedFor().isBefore(LocalDateTime.now())) {
+            throw new TaskExpiredException("Time for finish task with id " + checkedTask.getId() + " had expired at: " + checkedTask.getPlannedFor());
         } else {
             checkedTask.setImportance(task.getImportance());
             checkedTask.setDescription(task.getDescription());
-            checkedTask.setDone(task.isDone());
+            checkedTask.setDone(task.getDone());
             return taskRepository.save(checkedTask);
         }
     }
